prepared sessionlist to use server state
