+++ client/src/SessionsList/getNewListIdentifier.js
@@ -1,0 +1,3 @@
+export async function getNewListIdentifier() {
+  return 'TESTID'
+}
