+++ client/src/SessionsList/index.js
@@ -1,0 +1,1 @@
+export { SessionsListContainer } from './SessionsListContainer'
